Tuto Added
Tuto has been added to the game and animated for walking and running. Jumping still poses a challenge and there are a couple bugs. 1.) On startup the camera flips from front to back real quick to adjust to Tuto. 2.) When moving the camera quickly there is a flash on the screen where it gets split between scenes. May be a refresh rate thing.