Forest Level
The forest scene is completed. It also has portal animation for the ending, and a script for the portal when the player(Tuto)'s group hits the collision box.

The portal is also standalone, which means there are no more pillars around it. It can be used to make your own portal.