Maze and main scene update
Walls in main scene can no longer "clip" when the player runs by them but now the walls may be seen through at certain angles. In the main scene portals have been adjusted to no longer be accessible visually from the back. In addition to this boundaries have been added to the scene to stop players from breaking out of the world.