UI update (text and arrows)
When pressing f on any level an arrow will appear in front of the player pointing toward the exit. In addition the on screen text now updates for the forest level when they try to exit pre slime defeat, on main with each slime killed, and portal area entered, and on the maze level if they exit without the key.